completed enemy revisions
reworked enemies for more complexity and elegancy of function.   Created
enemyinfo.lua for enemy reference to be recorded and referenced.  In
enemyinfo the function refEnemy(typeEn) is defined, it returns data
about enemies of a specified type.  Updated enemies.lua for
compatability, still need to complete revisions on bullets.lua for
compatability.